Refactor the rest of the precompiles
