+++ packages/evm/src/ethers/governancePrecompile.ts
@@ -1,5 +1,5 @@
-import { ContractRunner, ethers, InterfaceAbi } from 'ethers';
-import { Abi } from 'viem';
+import { BigNumberish, Contract, ContractRunner, InterfaceAbi } from 'ethers';
+import { GOVERNANCE_PRECOMPILE_ABI, GOVERNANCE_PRECOMPILE_ADDRESS } from '../precompiles';
 
 /**
  * Represents the functions available in the Governance precompile contract,
@@ -13,7 +13,7 @@ export interface GovernancePrecompileFunctions {
 	 * @returns A Promise resolving to an object indicating the success of the transaction.
 	 * @category Cosmos Interoperability
 	 */
-	deposit(proposalID: ethers.BigNumberish): Promise<{ success: boolean }>;
+	deposit(proposalID: BigNumberish): Promise<{ success: boolean }>;
 	/**
 	 * Votes on a governance proposal.
 	 * @param proposalID The ID of the proposal to vote on.
@@ -21,86 +21,13 @@ export interface GovernancePrecompileFunctions {
 	 * @returns A Promise resolving to an object indicating the success of the transaction.
 	 * @category Cosmos Interoperability
 	 */
-	vote(proposalID: ethers.BigNumberish, option: ethers.BigNumberish): Promise<{ success: boolean }>;
+	vote(proposalID: BigNumberish, option: BigNumberish): Promise<{ success: boolean }>;
 }
 
 /** Represents the typed contract instance for the GOVERNANCE precompile contract.
  * @category Cosmos Interoperability
  * */
-export type GovernancePrecompileContract = ethers.Contract & GovernancePrecompileFunctions;
-
-/**
- * The address of the GOVERNANCE precompile contract, which can be used for interoperability between the EVM and Cosmos.
- *
- * @example
- * Wagmi
- * ```tsx
- * import { GOVERNANCE_PRECOMPILE_ADDRESS, GOVERNANCE_PRECOMPILE_ABI } from '@sei-js/evm';
- * import { useReadContract } from 'wagmi';
- **
- * const depositResponse = useContractFunction(
- *   GOVERNANCE_PRECOMPILE_ADDRESS,
- *   GOVERNANCE_PRECOMPILE_ABI,
- *   'deposit'
- * );
- * ```
- *
- * @example
- * ethers v6
- * ```tsx
- * import { getGovernancePrecompileEthersV6Contract, GOVERNANCE_PRECOMPILE_ADDRESS, parseSei } from '@sei-js/evm';
- * import { ethers } from 'ethers';
- *
- * const provider = new ethers.BrowserProvider(); // or any other provider
- * const signer = provider.getSigner();
- *
- * const governancePrecompileContract = getGovernancePrecompileEthersV6Contract(GOVERNANCE_PRECOMPILE_ADDRESS, signer);
- *
- * //Surround with try/catch for detailed errors
- * const depositResponse = await governancePrecompileContract.connect(signer).deposit('1', { value: parseSei(1) });
- * ```
- *
- * @category Cosmos Interoperability
- * */
-export const GOVERNANCE_PRECOMPILE_ADDRESS: `0x${string}` = '0x0000000000000000000000000000000000001006';
-
-/**
- * The ABI for the governance precompile contract, which can be used for interoperability between the EVM and Cosmos.
- *
- * @example
- * Wagmi
- * ```tsx
- * import { GOVERNANCE_PRECOMPILE_ADDRESS, GOVERNANCE_PRECOMPILE_ABI } from '@sei-js/evm';
- * import { useContractFunction } from '@wagmi/core';
- *
- * const depositResponse = useContractFunction(
- *   GOVERNANCE_PRECOMPILE_ADDRESS,
- *   GOVERNANCE_PRECOMPILE_ABI,
- *   'deposit'
- * );
- * ```
- *
- * @category Cosmos Interoperability
- * */
-export const GOVERNANCE_PRECOMPILE_ABI: Abi = [
-	{
-		inputs: [{ internalType: 'uint64', name: 'proposalID', type: 'uint64' }],
-		name: 'deposit',
-		outputs: [{ internalType: 'bool', name: 'success', type: 'bool' }],
-		stateMutability: 'payable',
-		type: 'function'
-	},
-	{
-		inputs: [
-			{ internalType: 'uint64', name: 'proposalID', type: 'uint64' },
-			{ internalType: 'int32', name: 'option', type: 'int32' }
-		],
-		name: 'vote',
-		outputs: [{ internalType: 'bool', name: 'success', type: 'bool' }],
-		stateMutability: 'nonpayable',
-		type: 'function'
-	}
-];
+export type GovernancePrecompileContract = Contract & GovernancePrecompileFunctions;
 
 /**
  * The ABI for the governance precompile contract, which can be used for interoperability between the EVM and Cosmos.
@@ -145,5 +72,5 @@ export const ETHERS_GOVERNANCE_PRECOMPILE_ABI = GOVERNANCE_PRECOMPILE_ABI as Int
  * @category Cosmos Interoperability
  */
 export const getGovernancePrecompileEthersV6Contract = (runner: ContractRunner) => {
-	return new ethers.Contract(GOVERNANCE_PRECOMPILE_ADDRESS, ETHERS_GOVERNANCE_PRECOMPILE_ABI, runner) as GovernancePrecompileContract;
+	return new Contract(GOVERNANCE_PRECOMPILE_ADDRESS, ETHERS_GOVERNANCE_PRECOMPILE_ABI, runner) as GovernancePrecompileContract;
 };
