--- packages/evm/src/oldPrecompiles/index.ts
@@ -1,10 +1,0 @@
-export * from './bank';
-export * from './distribution';
-export * from './governance';
-export * from './ibc';
-export * from './json';
-export * from './oracle';
-export * from './pointer';
-export * from './pointerview';
-export * from './staking';
-export * from './wasm';
