+++ packages/evm/src/precompiles/index.ts
@@ -1,1 +1,11 @@
 export * from './address';
+export * from './bank';
+export * from './distribution';
+export * from './governance';
+export * from './ibc';
+export * from './json';
+export * from './oracle';
+export * from './pointer';
+export * from './pointerview';
+export * from './staking';
+export * from './wasm';
