+++ packages/evm/src/viem/addressPrecompile.ts
@@ -1,4 +1,9 @@
 import { Abi } from 'viem';
 import { ADDRESS_PRECOMPILE_ABI } from '../precompiles';
 
+/**
+ * The Viem ABI for the bank precompile contract.
+ *
+ * @category Cosmos Interoperability
+ */
 export const VIEM_ADDRESS_PRECOMPILE_ABI = ADDRESS_PRECOMPILE_ABI as Abi;
