+++ packages/evm/src/viem/distributionPrecompile.ts
@@ -1,0 +1,9 @@
+import { Abi } from 'viem';
+import { DISTRIBUTION_PRECOMPILE_ABI } from '../precompiles';
+
+/**
+ * The Viem ABI for the distribution precompile contract.
+ *
+ * @category Cosmos Interoperability
+ */
+export const VIEM_DISTRIBUTION_PRECOMPILE_ABI = DISTRIBUTION_PRECOMPILE_ABI as Abi;
