+++ packages/evm/src/viem/ibcPrecompile.ts
@@ -1,0 +1,9 @@
+import { Abi } from 'viem';
+import { IBC_PRECOMPILE_ABI } from '../precompiles';
+
+/**
+ * The Viem ABI for the IBC precompile contract.
+ *
+ * @category Cosmos Interoperability
+ */
+export const VIEM_IBC_PRECOMPILE_ABI = IBC_PRECOMPILE_ABI as Abi;
