+++ packages/evm/src/viem/index.ts
@@ -1,1 +1,11 @@
 export * from './addressPrecompile';
+export * from './bankPrecompile';
+export * from './distributionPrecompile';
+export * from './governancePrecompile';
+export * from './ibcPrecompile';
+export * from './jsonPrecompile';
+export * from './oraclePrecompile';
+export * from './pointerPrecompile';
+export * from './pointerviewPrecompile';
+export * from './stakingPrecompile';
+export * from './wasmPrecompile';
