+++ packages/evm/src/viem/pointerPrecompile.ts
@@ -1,0 +1,9 @@
+import { Abi } from 'viem';
+import { POINTER_PRECOMPILE_ABI } from '../precompiles';
+
+/**
+ * The Viem ABI for the Pointer precompile contract.
+ *
+ * @category Cosmos Interoperability
+ */
+export const VIEM_POINTER_PRECOMPILE_ABI = POINTER_PRECOMPILE_ABI as Abi;
