+++ packages/evm/src/viem/pointerviewPrecompile.ts
@@ -1,0 +1,9 @@
+import { Abi } from 'viem';
+import { POINTERVIEW_PRECOMPILE_ABI } from '../precompiles';
+
+/**
+ * The Viem ABI for the Pointerview precompile contract.
+ *
+ * @category Cosmos Interoperability
+ */
+export const VIEM_POINTERVIEW_PRECOMPILE_ABI = POINTERVIEW_PRECOMPILE_ABI as Abi;
